Refactor top bar
