+++ src/App.tsx
@@ -13,6 +13,7 @@ import { openModal } from './store/GlobalModal';
 import { changeUsername } from './store/User';
 import APIAdapter from './services/api';
 import { StoreState } from './store';
+import TopBar from './components/TopBar';
 
 const App: React.FC = () => {
   const dispatch = useDispatch();
@@ -59,7 +60,14 @@ const App: React.FC = () => {
 
   return (
     <>
-      {isLogged || hasToken() ? <Routes /> : <BeforeLoginRoutes />}
+      {isLogged || hasToken() ? (
+        <>
+          <TopBar />
+          <Routes />
+        </>
+      ) : (
+        <BeforeLoginRoutes />
+      )}
       <GlobalStyle />
       <GlobalModal />
     </>
