+++ src/screens/EditUser/index.tsx
@@ -10,8 +10,6 @@ import {
   ProfileImageContainer,
 } from './styles';
 
-import { authenticationSuccessHandler } from '../../services/auth';
-import TopBar from '../TopBar';
 import APIAdapter from '../../services/api';
 import TextInput from '../../components/TextInput';
 import {
@@ -84,53 +82,50 @@ const EditUser: React.FC = () => {
   };
 
   return (
-    <>
-      <TopBar />
-      <Container>
-        <ProfileImageContainer>
-          <ProfileImage>A</ProfileImage>
+    <Container>
+      <ProfileImageContainer>
+        <ProfileImage>A</ProfileImage>
 
-          <MdAddAPhoto size="32" color="var(--white)" />
-        </ProfileImageContainer>
+        <MdAddAPhoto size="32" color="var(--white)" />
+      </ProfileImageContainer>
 
-        <FormContainer>
-          <TextInput
-            value={name}
-            disabled={loading}
-            variant="outlined"
-            label="Nome Completo"
-            placeholder="Digite o nome"
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-              setName(e.target.value)
-            }
-          />
+      <FormContainer>
+        <TextInput
+          value={name}
+          disabled={loading}
+          variant="outlined"
+          label="Nome Completo"
+          placeholder="Digite o nome"
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+            setName(e.target.value)
+          }
+        />
 
-          <TextInput
-            value={nickname}
-            disabled={loading}
-            variant="outlined"
-            label="Nome de usuário (Apelido)"
-            placeholder="Digite o nome de usuário"
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-              setNickname(e.target.value)
-            }
-          />
+        <TextInput
+          value={nickname}
+          disabled={loading}
+          variant="outlined"
+          label="Nome de usuário (Apelido)"
+          placeholder="Digite o nome de usuário"
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+            setNickname(e.target.value)
+          }
+        />
 
-          <TextInput
-            value={phone}
-            label="Telefone"
-            disabled={loading}
-            variant="outlined"
-            placeholder="Digite o telefone"
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-              setPhone(e.target.value)
-            }
-          />
+        <TextInput
+          value={phone}
+          label="Telefone"
+          disabled={loading}
+          variant="outlined"
+          placeholder="Digite o telefone"
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+            setPhone(e.target.value)
+          }
+        />
 
-          <SubmitBtn onClick={send}>CONFIRMAR</SubmitBtn>
-        </FormContainer>
-      </Container>
-    </>
+        <SubmitBtn onClick={send}>CONFIRMAR</SubmitBtn>
+      </FormContainer>
+    </Container>
   );
 };
