+++ src/screens/Offer/Create/index.tsx
@@ -9,7 +9,7 @@ import {
   openRequestSuccessModal,
 } from '../../../utils/requestModal';
 import { compressImages, parsePicturesToBase64 } from '../../../utils/images';
-import TopBar from '../../TopBar';
+import TopBar from '../../../components/TopBar';
 import { Container } from './styles';
 
 const isStrInvalid = (value: string | null | undefined) => !value;
@@ -123,12 +123,9 @@ const CreateOffer: React.FC = () => {
   };
 
   return (
-    <>
-      <TopBar />
-      <Container>
-        <GameForm loading={loading} handleSubmit={createOffer} />
-      </Container>
-    </>
+    <Container>
+      <GameForm loading={loading} handleSubmit={createOffer} />
+    </Container>
   );
 };
